:sparkles: api: move findAllCalibratedChallenges to CalibratedChallenge model
